feat: refactor WorkList component and integrate task fetching
- Updated App.vue to import WorkList from views instead of components.
- Modified WorkList.vue to create a new view for displaying tasks.
- Implemented task fetching from the server based on employeeId prop.
- Added automatic refresh of tasks every minute.
- Created a simple layout with a task list and main content area.
- Added a PowerShell script for testing the login API.
