+++ application-dev.yml
@@ -1,0 +1,7 @@
+minilab:
+  datasource:
+    driver-class-name: jdbc:mysql://minilab.mysql.database.azure.com:3306
+    host: localhost
+    port: 3306
+    username: Lu
+    password: Welly114514
